backend/requirements.txt
fastapi
uvicorn
openai
qdrant-client
sentence-transformers
torch
asyncpg
python-dotenv
beautifulsoup4
requests
pydantic-settings
google-generativeai